updated Makefile for tests
