+++ bootstrap/pkg/apis/apps/group.go
@@ -15,6 +15,20 @@
 // Package apps contains apps API versions
 package apps
 
+import (
+	"fmt"
+	log "github.com/sirupsen/logrus"
+	"io"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
+	"os"
+	"path/filepath"
+	"regexp"
+	"strings"
+)
+
 const (
 	DefaultNamespace = "kubeflow"
 	DefaultPlatform  = "none"
@@ -33,4 +47,106 @@ type KfApp interface {
 	Delete() error
 	Generate() error
 	Init() error
+}
+
+func KubeConfigPath() string {
+	kubeconfigEnv := os.Getenv("KUBECONFIG")
+	if kubeconfigEnv == "" {
+		home := os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
+		if home == "" {
+			for _, h := range []string{"HOME", "USERPROFILE"} {
+				if home = os.Getenv(h); home != "" {
+					break
+				}
+			}
+		}
+		kubeconfigPath := filepath.Join(home, ".kube", "config")
+		return kubeconfigPath
+	}
+	return kubeconfigEnv
+}
+
+// BuildOutOfClusterConfig returns k8s config
+func BuildOutOfClusterConfig() (*rest.Config, error) {
+	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
+	loadingRules.ExplicitPath = KubeConfigPath()
+	config, err := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
+		loadingRules, &clientcmd.ConfigOverrides{}).ClientConfig()
+	if err != nil {
+		return nil, err
+	}
+	return config, nil
+}
+
+func ServerVersion() (host string, version string, err error) {
+	restApi, err := BuildOutOfClusterConfig()
+	if err != nil {
+		return "", "", fmt.Errorf("couldn't build out-of-cluster config. Error: %v", err)
+	}
+	clnt, clntErr := kubernetes.NewForConfig(restApi)
+	if clntErr != nil {
+		return "", "", fmt.Errorf("couldn't get clientset. Error: %v", err)
+	}
+	serverVersion, serverVersionErr := clnt.ServerVersion()
+	if serverVersionErr != nil {
+		return "", "", fmt.Errorf("couldn't get server version info. Error: %v", serverVersionErr)
+	}
+	re := regexp.MustCompile("^v[0-9]+.[0-9]+.[0-9]+")
+	version = re.FindString(serverVersion.String())
+	return restApi.Host, "version:" + version, nil
+}
+
+func GetClientConfig() (*clientcmdapi.Config, error) {
+	kubeconfig := KubeConfigPath()
+	config, configErr := clientcmd.LoadFromFile(kubeconfig)
+	if configErr != nil {
+		return nil, fmt.Errorf("could not load config Error: %v", configErr)
+
+	}
+	return config, nil
+}
+
+// GetClientOutOfCluster returns a k8s clientset to the request from outside of cluster
+func GetClientOutOfCluster() (kubernetes.Interface, error) {
+	config, err := BuildOutOfClusterConfig()
+	if err != nil {
+		log.Fatalf("Can not get kubernetes config: %v", err)
+	}
+
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		log.Fatalf("Can not get kubernetes client: %v", err)
+	}
+
+	return clientset, nil
+}
+
+// capture replaces os.Stdout with a writer that buffers any data written
+// to os.Stdout. Call the returned function to cleanup and get the data
+// as a string.
+func capture() func() (string, error) {
+	r, w, err := os.Pipe()
+	if err != nil {
+		panic(err)
+	}
+
+	done := make(chan error, 1)
+
+	save := os.Stdout
+	os.Stdout = w
+
+	var buf strings.Builder
+
+	go func() {
+		_, err := io.Copy(&buf, r)
+		_ = r.Close()
+		done <- err
+	}()
+
+	return func() (string, error) {
+		os.Stdout = save
+		_ = w.Close()
+		err := <-done
+		return buf.String(), err
+	}
 }
