+++ bootstrap/pkg/apis/apps/ksapp/v1alpha1/register.go
@@ -14,12 +14,12 @@
 
 // NOTE: Boilerplate only.  Ignore this file.
 
-// Package v1alpha1 contains API Schema definitions for the apps v1alpha1 API group
+// Package v1alpha1 contains API Schema definitions for the ksapp v1alpha1 API group
 // +k8s:openapi-gen=true
 // +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps
+// +k8s:conversion-gen=github.com/kubeflow/kubeflow/bootstrap/pkg/apis/apps/ksapp
 // +k8s:defaulter-gen=TypeMeta
-// +groupName=apps.kubeflow.org
+// +groupName=ksapp.apps.kubeflow.org
 package v1alpha1
 
 import (
